# social_graph_service/src/config/swaggerDoc.yaml
openapi: 3.0.1
info:
  title: moody Social Graph Service
  description: This is the REST API documentation for the moody social_graph_service
  version: 1.0.0
servers:
  - url: /api
...
paths:
...
  /messages:
    post:
      tags:
        - message
      summary: Create a new message
      description: A new message is created and linked to a profile.
      requestBody:
        description: The message I want to send.
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  example: My personal message :)
                emoji:
                  type: string
                  example: '🤢'
        required: true
      responses:
        201:
          description: Profile created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 201
                  message:
                    $ref: '#/components/schemas/Message'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
    get:
      tags:
        - message
      summary: Get all messages
      description: Fetching all existing messages.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        404:
          description: No profiles found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  msg:
                    type: string
                    example: No messages found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
    delete:
      tags:
        - message
      summary: Delete all messages
      description: all existing messages are deleted (for testing purpose only)
      responses:
        200:
          description: All profiles removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  msg:
                    type: string
                    example: All messages removed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
  /messages/my:
    get:
      tags:
        - message
      summary: Get my messages
      description: through the auth header, the user object will be submitted and the last 30 messages of the user will be returned
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        404:
          description: No messages found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  msg:
                    type: string
                    example: No messages found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
  /messages/iFollow:
    get:
      tags:
        - message
      summary: Get messages from profiles I follow
      description: through the auth header, the user object will be submitted and the last 100 messages of profiles the user follows will be returned
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        404:
          description: No messages found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  msg:
                    type: string
                    example: No messages found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
  /messages/{username}:
    get:
      tags:
        - message
      summary: Get messages from a specific user
      description: Fetching all messages posted by a specific user
      parameters:
        - in: path
          name: username
          schema:
            type: string
            example: Birgit
          required: true
          description: the username of a profile
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        404:
          description: No messages found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  msg:
                    type: string
                    example: No messages found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
  /search/messages:
    get:
      tags:
        - search
      summary: Get messages from a specific search term
      description: Search for messages of my followed profiles given a specific search term (text)
      parameters:
        - name: text
          in: query
          required: true
          schema:
            type: string
            example: bla
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 200
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 401
                  msg:
                    type: string
                    example: Unauthorized - Token is not valid
        404:
          description: No messages found
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 404
                  msg:
                    type: string
                    example: No messages found
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: number
                    example: 500
                  msg:
                    type: string
                    example: Internal Server Error
      security:
        - authentication: []
components:
  schemas:
...
    Message:
      required:
        - text
        - emoji
      type: object
      properties:
        text:
          type: string
          example: My first message
        emoji:
          type: string
          example: 🤢
        author:
          type: string
          example: Hugo
        timestamp:
          type: string
          format: date-time
          example: Thu Dec 03 2020 13:23:42 GMT+0000 (Coordinated Universal Time)
  securitySchemes:
    authentication:
      type: apiKey
      name: Authorization
      in: header